@@ -14,7 +14,7 @@
 	<name>MarioKart_Backend</name>
 	<description>MarioKart_Backend</description>
 	<properties>
-		<java.version>21</java.version>
+		<java.version>25</java.version>
 	</properties>
 	<dependencies>
 
@@ -128,6 +128,24 @@
 						</exclude>
 					</excludes>
 				</configuration>
+			</plugin>
+			<plugin>
+				<groupId>org.openrewrite.maven</groupId>
+				<artifactId>rewrite-maven-plugin</artifactId>
+				<version>6.26.0</version>
+				<configuration>
+				<exportDatatables>true</exportDatatables>
+				<activeRecipes>
+					<recipe>org.openrewrite.java.migrate.UpgradeToJava25</recipe>
+				</activeRecipes>
+				</configuration>
+				<dependencies>
+				<dependency>
+					<groupId>org.openrewrite.recipe</groupId>
+					<artifactId>rewrite-migrate-java</artifactId>
+					<version>3.24.0</version>
+				</dependency>
+				</dependencies>
 			</plugin>
 		</plugins>
 	</build>
